# internal/send_video/req_models.go
package send_video

type AudioRequest struct {
	FileLink    string  `json:"file_link" binding:"required,url"`
	Language    string  `json:"language" binding:"required"` // TODO: Add support for dubbing multiple languages at the same time
	EmailId     string  `json:"email" binding:"required"`
	AudioLength float32 `json:"audio_length" binding:"required"`
}

type UploadHeader struct {
	FileName string
	Size     int
}

type UploadStatus struct {
	Code   int    `json:"code,omitempty"`
	Status string `json:"status,omitempty"`
	Pct    *int   `json:"pct,omitempty"`
}